+++ src/main/kotlin/br/com/org/gameslist/model/SignaturePlan.kt
@@ -1,20 +1,22 @@
 package br.com.org.gameslist.model
 
+import java.math.BigDecimal
+
 class SignaturePlan(type: String,
                     val monthlyPayment: Double,
                     val includedGames: Int,
                     val reputationDiscount: Double
 ): Plan(type) {
 
-    override fun getRentValue(rent: Rent): Double {
+    override fun getRentValue(rent: Rent): BigDecimal {
         val totalMonthGames = rent.player.monthGame(rent.period.initialDate.monthValue).size+1
 
         return if (totalMonthGames <= includedGames) {
-            0.0
+            BigDecimal.valueOf(0.0)
         } else {
             var originalValue = super.getRentValue(rent)
             if(rent.player.avg > 8) {
-                originalValue -= originalValue * reputationDiscount
+                originalValue -= originalValue * BigDecimal.valueOf(reputationDiscount)
             }
             originalValue
         }
